% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv-funcs.R
\name{cv.vsm2s}
\alias{cv.vsm2s}
\title{Cross-validation for two-step smoothing}
\usage{
cv.vsm2s(Y, x, lsp.f = NULL, nfpc = NULL, n.folds = 5, seed = NULL,
  argvals = NULL, raw = FALSE, ...)
}
\arguments{
\item{Y}{\eqn{n \times L} matrix of functional responses}

\item{x}{\eqn{n}-dimensional predictor vector}

\item{lsp.f}{candidate log smoothing parameter in the function direction}

\item{nfpc}{number of functional principal components}

\item{n.folds}{number of folds}

\item{seed}{seed for random number generation}

\item{argvals}{function argument values}

\item{raw}{use raw (first-step) coefficients}

\item{...}{other arguments passed to \code{\link{vsm2s}}}
}
\description{
Cross-validation function for \code{\link{vsm2s}}. Useful for choosing optimal 
function-direction smoothing parameter.
}

