% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-funcs.R
\name{plot.tps.fd}
\alias{plot.tps.fd}
\title{Plot a tps.fd object}
\usage{
\method{plot}{tps.fd}(x, newx = sort(x$x), xlab = "", ylab = "",
  label.curves = FALSE, argvals = x$argvals, lty = 1,
  owncolor = colorRamp(blue2red(100)), ...)
}
\arguments{
\item{x}{object created by \code{\link{tps.fd}}}

\item{newx}{new predictor values (at which to predict)}

\item{xlab, ylab}{labels for x-axis and y-axis}

\item{label.curves}{whether to label the curves}

\item{argvals}{function argument values}

\item{lty}{line type}

\item{owncolor}{user-defined color scheme}

\item{...}{other arguments, passed to \code{\link[graphics]{matplot}}}
}
\description{
Plot function for objects created by \code{\link{tps.fd}}
}

