% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-funcs.R
\name{rainbow.plot}
\alias{rainbow.plot}
\title{Rainbow plot}
\usage{
rainbow.plot(Y, x, argvals = NULL, subset = 1:length(x),
  label.curves = FALSE, xlab = "", ylab = "", legend = TRUE,
  digit = 1, posx = c(0.9, 0.93), posy = c(0.05, 0.9), zval = NULL, ...)
}
\arguments{
\item{Y}{n x L matrix of functional responses}

\item{x}{n-dimensional predictor vector}

\item{argvals}{function argument values}

\item{subset}{subset of functional observations to plot}

\item{label.curves}{whether to label the curves}

\item{xlab, ylab}{labels for x-axis and y-axis}

\item{legend}{logical: Should a legend be added?}

\item{digit, posx, posy, zval}{passed to \code{\link[shape]{colorlegend}}}

\item{...}{other arguments, passed to \code{\link{matplot}}}
}
\description{
Rainbow plot for functional data, color-coded by a predictor
}

