% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsm2s.r
\name{vsm2s}
\alias{vsm2s}
\title{Varying-smoother model via two-step smoothing}
\usage{
vsm2s(Y, x, argvals = NULL, k.f = 30 + (basistype.f == "fourier"),
  k.x = 15, lsp.x = seq(-10, 20, length.out = 100), lsp.f = NULL,
  norder.x = 4, basistype.f = "bspline", q.f = 2, q.x = norder.x - 2,
  varest = FALSE, nfpc = NULL, nfpc.default = 10, pve.cutoff = 0.99)
}
\arguments{
\item{Y}{\eqn{n \times L} matrix of functional responses}

\item{x}{\eqn{n}-dimensional predictor vector}

\item{argvals}{function argument values}

\item{k.f}{dimension of function-direction basis}

\item{k.x}{dimension of predictor-direction basis}

\item{lsp.x, lsp.f}{candidate log smoothing parameters in the predictor and function direction}

\item{norder.x}{order of splines in the predictor direction}

\item{basistype.f}{function-direction basis type for the functional direction}

\item{q.x, q.f}{penalty degree for the predictor- and function-direction bases}

\item{varest}{logical: should variance be estimated?}

\item{nfpc}{number of functional principal components}

\item{nfpc.default}{default number of functional principal components}

\item{pve.cutoff}{proportion of variance explained; used to choose number of FPCs}
}
\value{
An object of class \code{vsm2s}, which is a list whose elements include
\item{tpcoef,xcoef,coef.raw,coef.smoothed}{several types of spline coeffients}
\item{fitted}{fitted values}
\item{pwdf}{pointwise degrees of freedom}
}
\description{
Fits a nonlinear function-on-scalar regression model by (1) smoothing in the 
"predictor direction" at each point along the function, then (2) smoothing the result in the 
"function direction" (along the function). Supports only a single predictor.
}
\references{
Reiss, P. T., Huang, L., Chen, H., and Colcombe, S. Varying-smoother models for 
functional responses. Available at \url{https://arxiv.org/abs/1412.0778}

Reiss, P. T., Huang, L., Wu, P.-S., Chen, H., and Colcombe, S. Pointwise influence matrices for
functional-response regression. Available at \url{https://works.bepress.com/phil_reiss/43/}
}

